requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1

# Extraction
snscrape==0.6.2.20230320
requests==2.32.3
feedparser==6.0.11

# NLP & ML
spacy==3.7.5
transformers==4.44.2
torch==2.4.1

# Geocoding
geopy==2.4.1

# Database - VERSION CORRIGÉE pour Airflow
psycopg2-binary==2.9.9
elasticsearch==8.15.1
sqlalchemy==1.4.54

# Orchestration
apache-airflow==2.10.2

# Dashboard
streamlit==1.38.0
folium==0.17.0
plotly==5.24.1

# API
fastapi==0.115.0
uvicorn==0.30.6

# Tests
pytest==8.3.3
pytest-cov==5.0.0
